UJS-453: Fixed stone weight.

+++ database/migrations/2018_01_10_113622_create_model_stones_table.php
@@ -18,7 +18,7 @@ class CreateModelStonesTable extends Migration
             $table->integer('model_id')->unsigned();
             $table->integer('stone_id')->unsigned();
             $table->integer('amount');
-            $table->integer('weight');
+            $table->decimal('weight', 10, 3);
             $table->enum('flow', ['yes', 'no'])->default('no');
             $table->timestamps();
         });
